requirements.txt
# Core Spark & Data Processing
pyspark==3.5.0
delta-spark==3.1.0

# Data Generation & Testing
faker==24.0.0
pytest==8.0.0
pytest-mock==3.12.0

# Environment & Config
python-dotenv==1.0.1
pyyaml==6.0.1

# Optional: For data visualization/inspection
pandas==2.2.0